refactor common
