+++ src/cs/Bootsharp.Publish/Pack/BindingGenerator/BindingGenerator.cs
@@ -29,6 +29,7 @@ internal sealed class BindingGenerator (Preferences prefs)
             .OrderBy(m => m.Namespace).ToArray();
         if (bindings.Length == 0) return "";
         EmitImports();
+        builder.Append("\n\n");
         if (inspection.InstancedInterfaces.Count > 0)
             builder.Append(classGenerator.Generate(inspection.InstancedInterfaces));
         for (index = 0; index < bindings.Length; index++)
@@ -36,16 +37,17 @@ internal sealed class BindingGenerator (Preferences prefs)
         return builder.ToString();
     }
 
-    private void EmitImports ()
-    {
-        builder.Append("import { exports } from \"./exports\";\n");
-        builder.Append("import { Event } from \"./event\";\n");
-        builder.Append("import { registerInstance, getInstance, disposeOnFinalize } from \"./instances\";\n\n");
-        builder.Append("function getExports () { if (exports == null) throw Error(\"Boot the runtime before invoking C# APIs.\"); return exports; }\n");
-        builder.Append("function serialize(obj) { return JSON.stringify(obj); }\n");
-        builder.Append("function deserialize(json) { const result = JSON.parse(json); if (result === null) return undefined; return result; }\n\n");
-        builder.Append("/* v8 ignore start */\n");
-    }
+    private void EmitImports () => builder.Append(
+        """
+        import { exports } from "./exports";
+        import { Event } from "./event";
+        import { registerInstance, getInstance, disposeOnFinalize } from "./instances";
+
+        function getExports() { if (exports == null) throw Error("Boot the runtime before invoking C# APIs."); return exports; }
+
+        /* v8 ignore start */
+        """
+    );
 
     private void EmitBinding ()
     {
